+++ src/App.jsx
@@ -14,6 +14,7 @@ import AircraftProfile from './pages/AircraftProfile.jsx'
 import LogbookAudit from './pages/LogbookAudit.jsx'
 import GearRigging from './pages/GearRigging.jsx'
 import Compliance from './pages/Compliance.jsx'
+import AdCompliancePage from './pages/AdCompliancePage.jsx'
 import Checklists from './pages/Checklists.jsx'
 import ClaimListing from './pages/ClaimListing.jsx'
 import ReportView from './pages/ReportView.jsx'
@@ -123,6 +124,14 @@ function App() {
         element={
           <ProtectedRoute>
             <Compliance />
+          </ProtectedRoute>
+        }
+      />
+      <Route
+        path="/app/inspections/:id/ad-compliance"
+        element={
+          <ProtectedRoute>
+            <AdCompliancePage />
           </ProtectedRoute>
         }
       />